featrue : #6 상품 생성 API에 대해 헥사고날 아키텍처 적용 후 인수테스트및 단위테스트를 통해 확인한다.
