+++ mini-commerce/src/main/java/me/hero/minicommerce/item/application/port/in/CreateItemUseCase.java
@@ -1,0 +1,7 @@
+package me.hero.minicommerce.item.application.port.in;
+
+import me.hero.minicommerce.item.application.port.in.dto.CreateItemDto;
+
+public interface CreateItemUseCase {
+  CreateItemDto createItem(CreateItemDto itemDto);
+}
