+++ mini-commerce/src/test/java/me/hero/minicommerce/item/service/ItemServiceTest.java
@@ -12,9 +12,9 @@ import java.util.Optional;
 import me.hero.minicommerce.item.application.ItemService;
 import me.hero.minicommerce.item.domain.Item;
 import me.hero.minicommerce.item.adapter.out.persistance.ItemRepository;
-import me.hero.minicommerce.item.application.dto.FindItemDto;
-import me.hero.minicommerce.item.application.dto.ModifiedItemDto;
-import me.hero.minicommerce.item.application.dto.ModifyItemDto;
+import me.hero.minicommerce.item.application.port.in.dto.FindItemDto;
+import me.hero.minicommerce.item.application.port.in.dto.ModifiedItemDto;
+import me.hero.minicommerce.item.application.port.in.dto.ModifyItemDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
